Crate kimchi_msm

Modules

columns

constraint

mvlookup

Implement the protocol MVLookup https://eprint.iacr.org/2022/1530.pdf

precomputed_srs

Clone of kimchi/precomputed_srs.rs but for MSM project with BN254

proof

prover

serialization

verifier

witness

Constants

DOMAIN_SIZE

Domain size for the MSM project, equal to the BN254 SRS size.

LIMBS_NUM

Number of limbs representing one foreign field element (either Ff1 or Ff2).

LIMB_BITSIZE

Bitsize of the foreign field limb representation.

N

Number of columns FIXME: we must move it into the subdirectory of the foreign field addition circuit

Type Aliases

BN254

BN254G1Affine

BN254G2Affine

BaseSponge

Ff1

The foreign field we are emulating (one of the two)

Ff2

Fp

The native field we are working with.

OpeningProof

ScalarSponge

SpongeParams