kimchi_msm::circuit_design

Module composition

Structs

ComposedMPrism

IdMPrism

Identity MPrism from any type T to itself.

SubEnvColumn

Sub environment with a lens that is mapping columns.

SubEnvLookup

Sub environment with a lens that is mapping lookup tables.

Traits

MPrism

MPrism allows one to Something like a Prism, but for Maybe and not just any Applicative.