kimchi_msm::fec::interpreter

Function ec_add_circuit

pub fn ec_add_circuit<F: PrimeField, Ff: PrimeField, Env: ColWriteCap<F, FECColumn> + LookupCap<F, FECColumn, LookupTable<Ff>>>(
    env: &mut Env,
    xp: Ff,
    yp: Ff,
    xq: Ff,
    yq: Ff,
) -> (Ff, Ff)

Creates a witness for adding two points, p and q, each represented as a pair of foreign field elements. Returns a point.

This function is witness-generation counterpart (called by the prover) of constrain_ec_addition – see the documentation of the latter.