kimchi_msm::lookups

Type Alias LookupWitness

pub type LookupWitness<F> = LogupWitness<F, LookupTableIDs>;

Represents a witness of one instance of the lookup argument of the MSM project

Aliased Type

pub struct LookupWitness<F> {
    pub f: Vec<Vec<Logup<F, LookupTableIDs>>>,
    pub m: Vec<Vec<F>>,
}

Fields
f: Vec<Vec<Logup<F, LookupTableIDs>>>

A list of functions/looked-up values. Invariant: for fixed lookup tables, the last value of the vector is the lookup table t. The lookup table values must have a negative sign. The values are represented as: [ [f_{1}(1), …, f_{1}(ω^(n-1)], [f_{2}(1), …, f_{2}(ω^(n-1)] … [f_{m}(1), …, f_{m}(ω^(n-1)] ]

m: Vec<Vec<F>>

The multiplicity polynomials; by convention, this is a vector of columns, corresponding to the tail of f. That is, m[last] ~ f[last].

Implementations

impl<F: FftField> LookupWitness<F>

pub fn random(domain: EvaluationDomains<F>) -> (LookupTableIDs, Self)

Generate a random number of correct lookups in the table RangeCheck16