kimchi_msm::serialization::interpreter

Trait HybridSerHelpers

pub trait HybridSerHelpers<F: PrimeField, CIx: ColumnIndexer<usize>, LT: LookupTableID> {
    // Required method
    fn bitmask_be(
        &mut self,
        x: &<Self as ColAccessCap<F, CIx>>::Variable,
        highest_bit: u32,
        lowest_bit: u32,
        position: CIx,
    ) -> Self::Variable
       where Self: ColAccessCap<F, CIx>;
}

Required Methods

fn bitmask_be( &mut self, x: &<Self as ColAccessCap<F, CIx>>::Variable, highest_bit: u32, lowest_bit: u32, position: CIx, ) -> Self::Variable
where Self: ColAccessCap<F, CIx>,

Returns the bits between [highest_bit, lowest_bit] of the variable x, and copy the result in the column position. The value x is expected to be encoded in big-endian

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.