Trait kimchi_msm::serialization::interpreter::InterpreterEnv

pub trait InterpreterEnv {
    type Position;
    type Variable;

    // Required method
    fn deserialize_field_element(&mut self);

    // Provided methods
    fn range_check15(&mut self, _value: &Self::Variable) { ... }
    fn range_check4(&mut self, _value: &Self::Variable) { ... }
    fn copy(&mut self, _position: Self::Position, _value: Self::Variable) { ... }
}

Required Associated Types

type Position

type Variable

Required Methods

fn deserialize_field_element(&mut self)

Deserialize the next field element given as input

Provided Methods

fn range_check15(&mut self, _value: &Self::Variable)

Check that the value is in the range [0, 2^15-1]

fn range_check4(&mut self, _value: &Self::Variable)

Check that the value is in the range [0, 2^4-1]

fn copy(&mut self, _position: Self::Position, _value: Self::Variable)

Copy the value value in the column position