kimchi_msm::test::test_circuit

Module interpreter

Functions

build_fixed_selectors

Fixed selectors for the test circuit.

constrain_addition

A constraint function for A + B - C that reads values from limbs A and B, and additionally returns resulting value in C.

constrain_lookups

constrain_multiplication

A constraint function for A * B - D that reads values from limbs A and B, and multiplicationally returns resulting value in D.

constrain_test_const

A constraint function for A * B - D that reads values from limbs A and B, and multiplication_constally returns resulting value in D.

constrain_test_fixed_sel

A constraint function for A_0 + B_0 - FIXED_SEL_1

constrain_test_fixed_sel_degree_7

A constraint function for A_0^7 + B_0 - FIXED_SEL_1

constrain_test_fixed_sel_degree_7_mul_witness

A constraint function for 3 * A_0^7 + B_0 * FIXED_SEL_3

constrain_test_fixed_sel_degree_7_with_constants

A constraint function for 3 * A_0^7 + 42 * B_0 - FIXED_SEL_1

lookups_circuit

test_addition

Circuit generator function for A + B - C, with D = 0.

test_const

Circuit generator function for A_0 * B_0 - const, with every other column = 0

test_fixed_sel

Circuit generator function for A_0 + B_0 - FIXED_SEL_1.

test_fixed_sel_degree_7

Circuit generator function for A_0^7 + B_0 - FIXED_SEL_1.

test_fixed_sel_degree_7_mul_witness

Circuit generator function for 3 * A_0^7 + B_0 * FIXED_SEL_3.

test_fixed_sel_degree_7_with_constants

Circuit generator function for 3 * A_0^7 + 42 * B_0 - FIXED_SEL_1.

test_multiplication

Circuit generator function for A * B - C, with D = 0.