Crate kimchi_msm

Re-exports

pub use logup::Logup;

pub use logup::LogupWitness;

pub use logup::LookupProof as LogupProof;

pub use logup::LookupTable as LogupTable;

pub use logup::LookupTableID as LogupTableID;

pub use logup::LookupTableID;

Modules

circuit_design

column_env

columns

expr

fec

ffa

logup

Implement a variant of the logarithmic derivative lookups based on the equations described in the paper “Multivariate lookups based on logarithmic derivatives”.

lookups

Instantiations of Logups for the MSM project Instantiate the Logup protocol for the MSM project.

precomputed_srs

Clone of kimchi/precomputed_srs.rs but for MSM project with BN254

proof

prover

serialization

test

verifier

witness

Constants

DOMAIN_SIZE

Domain size for the MSM project, equal to the BN254 SRS size.

LIMB_BITSIZE

Bitsize of the foreign field limb representation.

N_LIMBS

Number of limbs representing one foreign field element (either Ff1 or Ff2).

Type Definitions

BN254

BN254G1Affine

BN254G2Affine

BaseSponge

Ff1

The foreign field we are emulating (one of the two)

Ff2

Fp

The native field we are working with.

OpeningProof

ScalarSponge

SpongeParams