Module kimchi_msm::circuit_design::capabilities

Traits

ColAccessCap

Environment capability for accessing and reading columns. This is necessary for building constraints.

ColWriteCap

Environment capability similar to ColAccessCap but for /also writing/ columns. Used on the witness side.

DirectWitnessCap

A direct field access capability modelling an abstract witness builder. Not for constraint building.

HybridCopyCap

Capability for computing arithmetic functions and enforcing constraints simultaneously.

LookupCap

Capability for invoking table lookups.

MultiRowReadCap

Capability for reading and moving forward in a multirow fashion. Holds a “current” row that can be moved forward with next_row. The ColWriteCap and ColAccessCap reason in terms of current row. The two other methods can be used to read/write previous.

Functions

read_column_array

Write an array of values simultaneously.

write_column_array

Write an array of values simultaneously.

write_column_array_const

Write an array of /field/ values simultaneously.

write_column_const

Write a field element directly as a constant.