Trait kimchi_msm::circuit_design::capabilities::MultiRowReadCap

pub trait MultiRowReadCap<F: PrimeField, CIx: ColumnIndexer>where
    Self: ColWriteCap<F, CIx>,{
    // Required methods
    fn read_row_column(&mut self, row: usize, col: CIx) -> Self::Variable;
    fn next_row(&mut self);
    fn curr_row(&self) -> usize;
}

Capability for reading and moving forward in a multirow fashion. Holds a “current” row that can be moved forward with next_row. The ColWriteCap and ColAccessCap reason in terms of current row. The two other methods can be used to read/write previous.

Required Methods

fn read_row_column(&mut self, row: usize, col: CIx) -> Self::Variable

Read value from a (row,column) position.

fn next_row(&mut self)

Progresses to the next row.

fn curr_row(&self) -> usize

Returns the current row.