Trait kimchi_msm::serialization::interpreter::InterpreterEnv

pub trait InterpreterEnv {
    type Position;
    type Variable: Clone + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Mul<Self::Variable, Output = Self::Variable> + Debug;

    // Required methods
    fn copy(
        &mut self,
        x: &Self::Variable,
        position: Self::Position
    ) -> Self::Variable;
    fn get_column_for_kimchi_limb(j: usize) -> Self::Position;
    fn get_column_for_intermediate_limb(j: usize) -> Self::Position;
    fn get_column_for_msm_limb(j: usize) -> Self::Position;
    fn constant(value: u128) -> Self::Variable;
    fn bitmask_be(
        &mut self,
        x: &Self::Variable,
        highest_bit: u32,
        lowest_bit: u32,
        position: Self::Position
    ) -> Self::Variable;

    // Provided methods
    fn range_check15(&mut self, _value: &Self::Variable) { ... }
    fn range_check4(&mut self, _value: &Self::Variable) { ... }
}

Required Associated Types

type Position

type Variable: Clone + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Mul<Self::Variable, Output = Self::Variable> + Debug

Required Methods

fn copy( &mut self, x: &Self::Variable, position: Self::Position ) -> Self::Variable

fn get_column_for_kimchi_limb(j: usize) -> Self::Position

fn get_column_for_intermediate_limb(j: usize) -> Self::Position

fn get_column_for_msm_limb(j: usize) -> Self::Position

fn constant(value: u128) -> Self::Variable

fn bitmask_be( &mut self, x: &Self::Variable, highest_bit: u32, lowest_bit: u32, position: Self::Position ) -> Self::Variable

Extract the bits from the variable x between highest_bit and lowest_bit, and store the result in position. lowest_bit becomes the least-significant bit of the resulting value. The value x is expected to be encoded in big-endian

Provided Methods

fn range_check15(&mut self, _value: &Self::Variable)

Check that the value is in the range [0, 2^15-1]

fn range_check4(&mut self, _value: &Self::Variable)

Check that the value is in the range [0, 2^4-1]

Object Safety

This trait is not object safe.