Trait kimchi_msm::serialization::interpreter::InterpreterEnv

pub trait InterpreterEnv<F: PrimeField, Ff: PrimeField> {
    type Position;
    type Variable: Clone + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Mul<Self::Variable, Output = Self::Variable> + Neg<Output = Self::Variable> + From<u64> + Debug;

    // Required methods
    fn add_constraint(&mut self, cst: Self::Variable);
    fn copy(
        &mut self,
        x: &Self::Variable,
        position: Self::Position
    ) -> Self::Variable;
    fn read_column(&self, pos: Self::Position) -> Self::Variable;
    fn get_column(pos: SerializationColumn) -> Self::Position;
    fn range_check15(&mut self, _value: &Self::Variable);
    fn range_check_abs15bit(&mut self, value: &Self::Variable);
    fn range_check_abs4bit(&mut self, value: &Self::Variable);
    fn range_check_ff_highest(&mut self, value: &Self::Variable);
    fn range_check4(&mut self, _value: &Self::Variable);
    fn constant(value: F) -> Self::Variable;
    fn bitmask_be(
        &mut self,
        x: &Self::Variable,
        highest_bit: u32,
        lowest_bit: u32,
        position: Self::Position
    ) -> Self::Variable;

    // Provided method
    fn read_column_direct(&self, pos: SerializationColumn) -> Self::Variable { ... }
}

Required Associated Types

type Position

type Variable: Clone + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Mul<Self::Variable, Output = Self::Variable> + Neg<Output = Self::Variable> + From<u64> + Debug

Required Methods

fn add_constraint(&mut self, cst: Self::Variable)

fn copy( &mut self, x: &Self::Variable, position: Self::Position ) -> Self::Variable

fn read_column(&self, pos: Self::Position) -> Self::Variable

fn get_column(pos: SerializationColumn) -> Self::Position

fn range_check15(&mut self, _value: &Self::Variable)

Check that the value is in the range [0, 2^15-1]

fn range_check_abs15bit(&mut self, value: &Self::Variable)

Checks input |x| ∈ [0,2^15)

fn range_check_abs4bit(&mut self, value: &Self::Variable)

Checks input |x| ∈ [0,2^4)

fn range_check_ff_highest(&mut self, value: &Self::Variable)

Checks x ∈ [0, f >> 15*16)

fn range_check4(&mut self, _value: &Self::Variable)

Check that the value is in the range [0, 2^4-1]

fn constant(value: F) -> Self::Variable

fn bitmask_be( &mut self, x: &Self::Variable, highest_bit: u32, lowest_bit: u32, position: Self::Position ) -> Self::Variable

Extract the bits from the variable x between highest_bit and lowest_bit, and store the result in position. lowest_bit becomes the least-significant bit of the resulting value. The value x is expected to be encoded in big-endian

Provided Methods

fn read_column_direct(&self, pos: SerializationColumn) -> Self::Variable

Object Safety

This trait is not object safe.