Struct kimchi_msm::witness::Witness

pub struct Witness<const N_WIT: usize, T> {
    pub cols: Box<[T; N_WIT]>,
}

The witness columns used by a gate of the MSM circuits. It is generic over the number of columns, N_WIT, and the type of the witness, T. It is parametrized by a type T which can be either:

Vec<G::ScalarField> for the evaluations

PolyComm<G> for the commitments It can be used to represent the different subcircuits used by the project.

Fields
cols: Box<[T; N_WIT]>

A witness row is represented by an array of N witness columns When T is a vector, then the witness describes the rows of the circuit.

Implementations

impl<const N_WIT: usize, T> Witness<N_WIT, T>

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

impl<const N_WIT: usize, T: Zero + Clone> Witness<N_WIT, Vec<T>>

pub fn zero_vec(domain_size: usize) -> Self

pub fn to_pub_columns<const NPUB: usize>(&self) -> Witness<NPUB, Vec<T>>